+++ realwh/ch02/Assign.hs
@@ -1,0 +1,8 @@
+{-
+  file: ch02/Assign.hs
+
+  This file is not a module.
+-}
+
+x = 10
+x = 11