+++ realwh/ch02/RoundToEven.hs
@@ -1,0 +1,7 @@
+{-
+  file: ch02/RoundToEven.hs
+
+  This file is not a module
+-}
+
+isOdd n = mod n 2 == 1