+++ realwh/ch02/Take.hs
@@ -1,0 +1,8 @@
+{-
+  file: ch02/Take.hs
+
+  This file is not a module.
+-}
+
+take :: Int -> ([a] -> [a])
+take = undefined