+++ realwh/ch02/shortCircuit.hs
@@ -1,0 +1,7 @@
+{-
+  file: ch02/shortCircuit.hs
+
+  This file is not a module.
+-}
+
+newOr a b = if a then a else b