+++ realwh/ch03/add.hs
@@ -1,0 +1,8 @@
+{-
+  file: ch03/add.hs
+
+  `add.hs` isn't a module.
+-}
+
+add a b = a + b
+